fixing page routes and build failures
